requirements.txt
airspeed==0.5.16
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
colorama==0.4.4
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
future==0.18.2
guided-filter-pytorch==3.7.5
imageio==2.9.0
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.2
olefile==0.46
opencv-python==4.4.0.46
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.8
Pygments==2.7.3
pyparsing==2.4.7
scikit-image==0.18.1
scipy==1.5.4
six==1.15.0
tensorflow-estimator==2.3.0
tifffile==2020.12.8
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tqdm==4.54.1
traitlets==5.0.5
wcwidth==0.2.5
wincertstore==0.2
pympler
